Refactor header components and update imports; remove unused styles
