+++ ParqueInfantil/frontend/src/components/ActivityContainer.jsx
