+++ ParqueInfantil/frontend/src/components/headers/header-educator.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { FaHome, FaListAlt, FaUser, FaTasks, FaChartBar } from 'react-icons/fa';
-import './header-educator.css';
+
 
 const HeaderEducator = ({ username }) => {
     return (
