--- ParqueInfantil/frontend/src/styles/AboutUs.css
@@ -1,94 +1,0 @@
-/* AboutUs.css */
-
-body {
-    font-family: 'Arial', sans-serif;
-    margin: 0;
-    padding: 0;
-    background-color: #a0cd98;
-    color: #333;
-}
-
-.container {
-    width: 80%;
-    margin: 0 auto;
-    padding: 20px;
-}
-
-.header {
-    text-align: center;
-    padding: 50px 0;
-}
-
-.header h1 {
-    font-size: 3em;
-    margin: 0;
-    color: #2c3e50;
-}
-
-.header p {
-    font-size: 1.2em;
-    color: #7f8c8d;
-}
-
-.section {
-    margin: 50px 0;
-}
-
-.section h2 {
-    font-size: 2.5em;
-    color: #2980b9;
-    margin-bottom: 20px;
-}
-
-.section p {
-    font-size: 1.1em;
-    line-height: 1.6;
-    color: #34495e;
-}
-
-.team {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-}
-
-.team-member {
-    background-color: #ecf0f1;
-    border-radius: 10px;
-    padding: 20px;
-    margin: 10px;
-    width: 30%;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    text-align: center;
-}
-
-.team-member img {
-    border-radius: 50%;
-    width: 150px;
-    height: 150px;
-    object-fit: cover;
-    margin-bottom: 15px;
-}
-
-.team-member h3 {
-    font-size: 1.5em;
-    color: #2c3e50;
-    margin: 10px 0;
-}
-
-.team-member p {
-    font-size: 1em;
-    color: #7f8c8d;
-}
-
-.footer {
-    text-align: center;
-    padding: 20px 0;
-    background-color: #2c3e50;
-    color: #ecf0f1;
-}
-
-.footer p {
-    margin: 0;
-    font-size: 1em;
-}
