--- ParqueInfantil/frontend/src/styles/Home.css
@@ -1,93 +1,0 @@
-/* Home.css */
-
-body {
-    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji";
-    margin: 0;
-    padding: 0;
-    background-color: #f6f8fa;
-}
-
-.header {
-    background-color: #24292e;
-    padding: 16px;
-    color: white;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.header .logo {
-    font-size: 24px;
-    font-weight: bold;
-}
-
-.header .nav {
-    display: flex;
-    gap: 16px;
-}
-
-.header .nav a {
-    color: white;
-    text-decoration: none;
-    font-size: 16px;
-}
-
-.header .nav a:hover {
-    text-decoration: underline;
-}
-
-.main {
-    padding: 40px;
-    max-width: 1200px;
-    margin: 0 auto;
-}
-
-.main .title {
-    font-size: 32px;
-    font-weight: bold;
-    margin-bottom: 20px;
-}
-
-.main .description {
-    font-size: 20px;
-    color: #586069;
-    margin-bottom: 40px;
-}
-
-.main .content {
-    display: flex;
-    gap: 20px;
-    flex-wrap: wrap;
-}
-
-.card {
-    background-color: white;
-    border: 1px solid #e1e4e8;
-    border-radius: 6px;
-    padding: 20px;
-    width: calc(33.333% - 20px);
-    box-shadow: 0 1px 3px rgba(27, 31, 35, 0.12);
-}
-
-.card:hover {
-    box-shadow: 0 4px 6px rgba(27, 31, 35, 0.15);
-}
-
-.card .card-title {
-    font-size: 20px;
-    font-weight: bold;
-    margin-bottom: 10px;
-}
-
-.card .card-description {
-    font-size: 16px;
-    color: #586069;
-}
-
-.footer {
-    background-color: #24292e;
-    color: white;
-    padding: 20px;
-    text-align: center;
-    margin-top: 40px;
-}
