--- ParqueInfantil/frontend/src/styles/Login.css
@@ -1,65 +1,0 @@
-body {
-    font-family: 'Arial', sans-serif;
-    background: linear-gradient(to right, #ffecd2, #fcb69f);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    margin: 0;
-}
-
-.login-container {
-    background: #a8f571;
-    padding: 2rem;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    max-width: 400px;
-    width: 100%;
-    text-align: center;
-}
-
-.login-container h1 {
-    font-size: 2rem;
-    margin-bottom: 1.5rem;
-    color: #ff6f61;
-}
-
-.login-container input[type="text"],
-.login-container input[type="password"] {
-    width: 100%;
-    padding: 0.75rem;
-    margin: 0.5rem 0;
-    border: 1px solid #ddd;
-    border-radius: 5px;
-    box-sizing: border-box;
-}
-
-.login-container button {
-    background: #ff6f61;
-    color: #fff;
-    padding: 0.75rem;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-    width: 100%;
-    font-size: 1rem;
-    transition: background 0.3s ease;
-}
-
-.login-container button:hover {
-    background: #ff4a3d;
-}
-
-.login-container p {
-    margin-top: 1rem;
-    color: #666;
-}
-
-.login-container a {
-    color: #ff6f61;
-    text-decoration: none;
-}
-
-.login-container a:hover {
-    text-decoration: underline;
-}
