+++ ParqueInfantil/frontend/src/styles/reservation-container.css
@@ -1,0 +1,108 @@
+.reservation-container {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+    padding: 20px;
+    background-color: #f9f9f9;
+    border-radius: 10px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.reservation-item {
+    display: flex;
+    align-items: center;
+    gap: 15px;
+    padding: 15px;
+    background-color: #fff;
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    transition: transform 0.2s;
+}
+
+.reservation-item:hover {
+    transform: translateY(-5px);
+}
+
+.activity-thumbnail {
+    width: 80px;
+    height: 80px;
+    border-radius: 8px;
+    object-fit: cover;
+}
+
+.reservation-info {
+    flex-grow: 1;
+}
+
+.reservation-info h3 {
+    margin: 0;
+    font-size: 1.2em;
+    color: #333;
+}
+
+.reservation-status {
+    margin: 5px 0 0;
+    font-size: 0.9em;
+    font-weight: bold;
+}
+
+.reservation-status.pending {
+    color: #ff9800;
+}
+
+.reservation-status.confirmed {
+    color: #4caf50;
+}
+
+.reservation-status.cancelled {
+    color: #f44336;
+}
+
+.reservation-actions {
+    display: flex;
+    gap: 10px;
+}
+
+.action-icon {
+    font-size: 1.5em;
+    cursor: pointer;
+    color: #888;
+    transition: color 0.2s;
+}
+
+.action-icon:hover {
+    color: #333;
+}
+
+.reservation-details {
+    padding: 20px;
+    background-color: #fff;
+    border-radius: 8px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.reservation-details h2 {
+    margin-top: 0;
+    font-size: 1.5em;
+    color: #333;
+}
+
+.reservation-details p {
+    margin: 10px 0;
+    font-size: 1em;
+    color: #666;
+}
+
+.reservation-details button {
+    padding: 10px 20px;
+    background-color: #4caf50;
+    color: #fff;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+    transition: background-color 0.2s;
+}
+
+.reservation-details button:hover {
+    background-color: #45a049;
+}
